# Add IPC stream interface for zero-copy Arrow data access